requirements.txt
discord.py>=2.3.0
PyYAML>=6.0
GitPython>=3.1.0
python-dotenv>=1.0.0
